Avoid module inception
